+++ apps/chevron/src/app/(survey)/(intake)/lexTale/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function page() {
+  return (
+    <div>page</div>
+  )
+}
+
+export default page
